+++ apps/dev/src/registries.tsx
@@ -137,6 +137,7 @@ export const gameRegistry: GameRegistry = {
   "floating-text": () => import("./demo/floatingTextDemo").then((module) => module.floatingTextDemoGame),
   "coach-marks": () => import("./demo/coachMarksDemo").then((module) => module.coachMarksDemoGame),
   "objective-banner": () => import("./demo/objectiveBannerDemo").then((module) => module.objectiveBannerDemoGame),
+  "key-rebinding": () => import("./demo/keyRebindingDemo").then((module) => module.keyRebindingDemoGame),
   "shop-grid": () => import("./demo/shopGridDemo").then((module) => module.shopGridDemoGame),
   "save-slots": () => import("./demo/saveSlotsDemo").then((module) => module.saveSlotsDemoGame),
   "kill-feed": () => import("./demo/killFeedDemo").then((module) => module.killFeedDemoGame),